added CompanyFragment, CompanyViewModel

+++ app/src/main/java/com/georgcantor/storetest/model/data/Company.kt
@@ -5,8 +5,8 @@ data class Company(
     var name: String,
     var img: String,
     var description: String,
-    var lat: Int,
-    var lon: Int,
+    var lat: Float,
+    var lon: Float,
     var www: String,
     var phone: String
 )
